+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Defects Corrected
  * Cleaned up dependencies and fixed few minor issues with generated code in archetype 
 
+### Enhancements
+ * Added toString implementation to DelegatingGenericResponse and BuiltGenericResponse
+
 ## 2.4 - 16 Feb 2017
 
 ### Additions